confirmation message before delete

+++ vue-app/src/views/JobView.vue
@@ -141,9 +141,12 @@ const state = reactive({
 
 const deleteJob = async () =>{
     try{
-        await axios.delete(`/api/jobs/${jobId}`);
-        toast.success('Jon deleted successfully');
-        router.push('/jobs');
+        const confirm = window.confirm('Are you sure you want to delete this job?');
+        if(confirm){
+            await axios.delete(`/api/jobs/${jobId}`);
+            toast.success('Jon deleted successfully');
+            router.push('/jobs');
+        }
     }catch(err){
         console.log('Error deleting job', err);
         toast.error('Job not deleted');
